backend/requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==25.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
